+++ aio_throttle/__init__.py
@@ -1,2 +1,3 @@
-from .throttle import ThrottlePriority, ThrottleStats, ThrottleResult, Throttler  # noqa
+from .throttle import Throttler  # noqa
 from .quotas import ThrottleCapacityQuota, MaxFractionCapacityQuota, ThrottleQuota, RandomRejectThrottleQuota  # noqa
+from .abc import ThrottlerBase, ThrottlePriority, ThrottleStats, ThrottleResult  # noqa
