+++ requirements-dev.txt
@@ -5,6 +5,7 @@ flake8==3.8.4
 mypy==0.790
 black==19.10b0
 aiohttp==3.6.2
+prometheus_client==0.7.1
 setuptools==50.3.2
 wheel==0.35.1
 twine==3.2.0
